+++ src/p700_799/e738.rs
@@ -1,6 +1,6 @@
 use crate::utils::{FIArray::FIArrayI64, multiplicative_function_summation::dirichlet_inv_i64};
 
-const N: i64 = 10;
+const N: usize = 10;
 const MOD: i64 = 1e9 as i64 + 7;
 
 pub fn main() {
